verify values being made against expected types
Add a function that confirms a composite component value that is
constructed from component value parts conforms to the defined type.

This verify step used to be performed by the wasmtime library when a
component type was asked to build a component value. The v20.0.0 library
was changed to have a simpler type system on the surface as the
`Resource` handling would require a `store` from the embedder.

So far, this WAVE library punts on trying to handle any `Resource`
variable so the checks made between variables and types are still
straight forward and no `store` is required.